+++ pycantonese/data/rime_cantonese/__init__.py
@@ -1,1 +1,17 @@
-# TODO: Functionality to read the JSON files.
+import json
+import os
+
+
+_THIS_DIR = os.path.abspath(os.path.dirname(__file__))
+
+with open(os.path.join(_THIS_DIR, "chars_to_jyutping.json")) as f:
+    CHARS_TO_JYUTPING = json.load(f)
+
+with open(os.path.join(_THIS_DIR, "lettered.json")) as f:
+    LETTERED = json.load(f)
+
+with open(os.path.join(_THIS_DIR, "phrases.json")) as f:
+    PHRASES = json.load(f)
+
+with open(os.path.join(_THIS_DIR, "maps.json")) as f:
+    MAPS = json.load(f)